requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
whitenoise==6.7.0
dj-database-url==2.2.0
python-decouple==3.8

psycopg2-binary==2.9.10

google-generativeai

torch==2.5.1
sentence-transformers==3.3.1
scikit-learn==1.6.0
numpy==2.2.0

twilio==9.3.5